Separacion de ProxyTests
